create class DAO

+++ src/feijuca/Produto.java
@@ -1,40 +1,19 @@
 package feijuca;
 
-/*
- * Copyright 2017 joao.palmas.br.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-
-/**
- *
- * @author pix01
- */
 public class Produto {
     private int codigo;
     private String nome;
-    private double valor;
+    private float valor;
 
     public Produto() {
     }
 
-    public Produto(int codigo, String nome, double valor) {
+    public Produto(int codigo, String nome, float valor) {
         this.codigo = codigo;
         this.nome = nome;
         this.valor = valor;
     }
-
+    
     public int getCodigo() {
         return codigo;
     }
@@ -51,11 +30,11 @@ public class Produto {
         this.nome = nome;
     }
 
-    public double getValor() {
+    public float getValor() {
         return valor;
     }
 
-    public void setValor(double valor) {
+    public void setValor(float valor) {
         this.valor = valor;
     }
